% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{ggboxplot}
\alias{ggboxplot}
\title{Make a boxplot using ggplot2}
\usage{
ggboxplot(data, x, y, cols, rows, line, color, show.points = TRUE,
  jitter = 0, box.size = 0.5, whisker.size = 0.5, whisker.width = 0.5,
  point.size = 3, point.alpha = 0.5, point.color = "black",
  color.pal = NULL, show.legend = TRUE)
}
\arguments{
\item{data}{A data.frame containing all the variables needed to draw the box
plot.}

\item{x}{A character string indicating the x axis variable.}

\item{y}{A character string indicating the y axis variable.}

\item{cols}{A character string indicating the variable defines faceting
groups on columns.}

\item{rows}{A character string indicating the variable defines faceting
groups on rows.}

\item{line}{A character string indicating the variable used to draw lines
between points.}

\item{color}{A character string indicating the variable defines the color of lines or points.}

\item{show.points}{Logic value indicates whether to show points.}

\item{jitter}{Numeric value defines the amount of horizontal jitter.}

\item{box.size}{Numeric value defines the size of the box.}

\item{whisker.size}{Numeric value defines the size of whisker.}

\item{whisker.width}{Numeric value defines the width of boxplot}

\item{point.size}{Numeric value defines the size of points.}

\item{point.alpha}{Numeric value defines the transparency.}

\item{point.color}{Character value defines the color of points when the
argument 'color' is not specified.}

\item{color.pal}{Character vector defines the color panel to use.}

\item{show.legend}{Logic value defines whether to show legend.}
}
\value{
A ggplot object
}
\description{
This is a encapsulated function to make a boxplot using the
ggplot2 package. When parsing a cross-over study or repeated measure data,
lines can be drawn between the 2 points belong to the same entity in
different timepoints.
}
\author{
Chenghao Zhu
}
