% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{ggscatterplot}
\alias{ggscatterplot}
\title{Make a scatterplot using ggplot2}
\usage{
ggscatterplot(data, x, y, cols, rows, color, trendline = TRUE,
  trendline.color = "steelblue", trendline.size = 1, show.points = TRUE,
  point.size = 3, point.alpha = 1, point.color = "black",
  color.pal = NULL, show.legend = TRUE)
}
\arguments{
\item{data}{A data.frame containing all the variables needed to draw the box
plot.}

\item{x}{A character string indicating the x axis variable.}

\item{y}{A character string indicating the y axis variable.}

\item{cols}{A character string indicating the variable defines faceting
groups on columns.}

\item{rows}{A character string indicating the variable defines faceting
groups on rows.}

\item{color}{A character string indicating the variable defines the color of
lines or points.}

\item{trendline}{Logic value whether to draw a linear trandline.}

\item{trendline.color}{DESCRIPTION.}

\item{trendline.size}{DESCRIPTION.}

\item{show.points}{DESCRIPTION.}

\item{point.size}{Numeric value defines the size of points.}

\item{point.alpha}{Numeric value defines the transparency.}

\item{point.color}{Character value defines the color of points when the
argument 'color' is not specified.}

\item{color.pal}{Character vector defines the color panel to use.}

\item{show.legend}{Logic value defines whether to show legend.}
}
\value{
a ggplot object
}
\description{
This is a encapsulated function to make a scatterplot using the
ggplot2 package. It allows the users to control the plot color, size, and
other styles using arguments.
}
\author{
Chenghao Zhu
}
